fix: move v0_8 inside src to fix CI symlink issues and support dual import paths
