+++ renderers/angular/src/v0_8/components/image.ts
