+++ renderers/angular/src/v0_8/components/multiple-choice.ts
