+++ renderers/angular/src/v0_8/components/text.ts
