+++ renderers/angular/src/v0_8/ng-package.json
