+++ renderers/angular/src/v0_8/rendering/catalog.ts
