+++ samples/client/angular/projects/lib/v0_8
@@ -1,0 +1,1 @@
+../../../../../renderers/angular/v0_8
